@@ -20,36 +20,36 @@
         "composer/package-versions-deprecated": "1.11.99.3",
         "doctrine/doctrine-bundle": "^2.3",
         "doctrine/doctrine-migrations-bundle": "^3.1",
-        "doctrine/orm": "^2.7",
+        "doctrine/orm": "^2.9",
         "knplabs/knp-menu-bundle": "^3.1",
-        "lexik/jwt-authentication-bundle": "^2.11",
+        "lexik/jwt-authentication-bundle": "^2.12",
         "sensio/framework-extra-bundle": "^6.1",
-        "sonata-project/admin-bundle": "^3.99",
-        "sonata-project/doctrine-orm-admin-bundle": "^3.34",
-        "symfony/asset": "4.4.*",
-        "symfony/console": "4.4.*",
-        "symfony/dotenv": "4.4.*",
-        "symfony/expression-language": "4.4.*",
-        "symfony/flex": "^1.12",
-        "symfony/form": "4.4.*",
-        "symfony/framework-bundle": "4.4.*",
-        "symfony/http-client": "4.4.*",
-        "symfony/mailer": "4.4.*",
+        "sonata-project/admin-bundle": "4.0.0-rc.2",
+        "sonata-project/doctrine-orm-admin-bundle": "4.0.0-rc.2",
+        "symfony/asset": "5.3.*",
+        "symfony/console": "5.3.*",
+        "symfony/dotenv": "5.3.*",
+        "symfony/expression-language": "5.3.*",
+        "symfony/flex": "^1.15",
+        "symfony/form": "5.3.*",
+        "symfony/framework-bundle": "5.3.*",
+        "symfony/http-client": "5.3.*",
+        "symfony/mailer": "5.3.*",
         "symfony/monolog-bundle": "^3.7",
-        "symfony/process": "4.4.*",
-        "symfony/proxy-manager-bridge": "4.4.*",
-        "symfony/security-bundle": "4.4.*",
-        "symfony/serializer": "4.4.*",
-        "symfony/translation": "4.4.*",
-        "symfony/twig-bundle": "^4.4",
-        "symfony/validator": "4.4.*",
-        "symfony/web-link": "4.4.*",
+        "symfony/process": "5.3.*",
+        "symfony/proxy-manager-bridge": "5.3.*",
+        "symfony/runtime": "5.3.*",
+        "symfony/security-bundle": "5.3.*",
+        "symfony/serializer": "5.3.*",
+        "symfony/translation": "5.3.*",
+        "symfony/twig-bundle": "^5.3",
+        "symfony/validator": "5.3.*",
+        "symfony/web-link": "5.3.*",
         "symfony/webpack-encore-bundle": "^1.11",
-        "symfony/yaml": "4.4.*",
-        "twig/extra-bundle": "^2.12|^3.3",
-        "twig/twig": "^2.12|^3.3",
-        "vyfony/filterable-table-bundle": "^0.5.5",
-        "vyfony/mysqldump-bundle": "^0.2.0"
+        "symfony/yaml": "5.3.*",
+        "twig/extra-bundle": "^3.3",
+        "twig/twig": "^3.3",
+        "vyfony/filterable-table-bundle": "v0.6.2"
     },
     "require-dev": {
         "enlightn/security-checker": "^1.9",
@@ -59,15 +59,14 @@
         "phpro/grumphp": "^1.3",
         "phpunit/phpunit": "^9.5",
         "squizlabs/php_codesniffer": "^3.6",
-        "symfony/browser-kit": "^4.4",
-        "symfony/css-selector": "^4.4",
-        "symfony/debug-bundle": "^4.4",
+        "symfony/browser-kit": "^5.3",
+        "symfony/css-selector": "^5.3",
+        "symfony/debug-bundle": "^5.3",
         "symfony/maker-bundle": "^1.31",
         "symfony/phpunit-bridge": "^5.2",
-        "symfony/stopwatch": "^4.4",
-        "symfony/var-dumper": "^4.4",
-        "symfony/web-profiler-bundle": "^4.4",
-        "symfony/web-server-bundle": "4.4.*"
+        "symfony/stopwatch": "^5.3",
+        "symfony/var-dumper": "^5.3",
+        "symfony/web-profiler-bundle": "^5.3"
     },
     "config": {
         "preferred-install": {
@@ -111,7 +110,7 @@
     "extra": {
         "symfony": {
             "allow-contrib": true,
-            "require": "4.4.*"
+            "require": "5.3.*"
         }
     }
 }
