@@ -36,6 +36,7 @@ final class AdminLoginForm extends AbstractType
     {
         $builder
             ->add('username', TextType::class)
-            ->add('password', PasswordType::class);
+            ->add('password', PasswordType::class)
+        ;
     }
 }
