@@ -27,14 +27,8 @@ namespace App\Repository\Media;
 
 use App\Entity\Media\File;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Doctrine\Common\Persistence\ManagerRegistry;
+use Doctrine\Persistence\ManagerRegistry;
 
-/**
- * @method File|null find(int $id, ?int $lockMode = null, ?int $lockVersion = null)
- * @method File|null findOneBy(array $criteria, ?array $orderBy = null)
- * @method File[]    findAll()
- * @method File[]    findBy(array $criteria, ?array $orderBy = null, ?int $limit = null, ?int $offset = null)
- */
 final class FileRepository extends ServiceEntityRepository
 {
     public function __construct(ManagerRegistry $registry)
